--- src/common/errors.js
@@ -1,7 +1,0 @@
-const errorMessage = {
-  500: "Internal"
-}
-
-module.exports = {
-    errorMessage
-}
