+++ src/common/utils.js
@@ -1,19 +1,3 @@
-function getPostData(req) {
-    return new Promise((resolve, reject) => {
-      try {
-        let body = '';
-        req.on('data', (chunk) => {
-          body += chunk.toString();
-        });
-        req.on('end', () => {
-          resolve(body);
-        });
-      } catch (error) {
-        reject(error);
-      }
-    });
-  }
-  
 const omitPassword = (user) => {
     const { password, ..._user } = user
     return _user
@@ -26,7 +10,6 @@ function mutationFilter(arr, cb) {
 }
 
   module.exports = {
-    getPostData,
     omitPassword,
     mutationFilter
   };
