+++ src/server.js
@@ -1,4 +1,3 @@
-
 const fastify = require('fastify')({
   logger: true
 })
